requirements.txt
faster-whisper
sounddevice
numpy
scipy
python-dotenv
